+++ include/wjr/x86/simd/simd_cast.hpp
@@ -94,14 +94,14 @@ struct simd_cast_fn<uint8_t, __m128i_t> {
 template <>
 struct simd_cast_fn<__m128i_t, int8_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int8_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si32(v);
+        return static_cast<int8_t>(_mm_cvtsi128_si32(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m128i_t, uint8_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint8_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si32(v);
+        return static_cast<uint8_t>(_mm_cvtsi128_si32(v));
     }
 };
 
@@ -122,14 +122,14 @@ struct simd_cast_fn<uint16_t, __m128i_t> {
 template <>
 struct simd_cast_fn<__m128i_t, int16_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int16_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si32(v);
+        return static_cast<int16_t>(_mm_cvtsi128_si32(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m128i_t, uint16_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint16_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si32(v);
+        return static_cast<uint16_t>(_mm_cvtsi128_si32(v));
     }
 };
 
@@ -157,7 +157,7 @@ struct simd_cast_fn<__m128i_t, int32_t> {
 template <>
 struct simd_cast_fn<__m128i_t, uint32_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint32_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si32(v);
+        return static_cast<uint32_t>(_mm_cvtsi128_si32(v));
     }
 };
 
@@ -171,7 +171,7 @@ struct simd_cast_fn<int64_t, __m128i_t> {
 template <>
 struct simd_cast_fn<uint64_t, __m128i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m128i operator()(uint64_t v) const {
-        return _mm_cvtsi64_si128(v);
+        return _mm_cvtsi64_si128(static_cast<int64_t>(v));
     }
 };
 
@@ -185,7 +185,7 @@ struct simd_cast_fn<__m128i_t, int64_t> {
 template <>
 struct simd_cast_fn<__m128i_t, uint64_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint64_t operator()(__m128i v) const {
-        return _mm_cvtsi128_si64(v);
+        return static_cast<uint64_t>(_mm_cvtsi128_si64(v));
     }
 };
 
@@ -264,63 +264,63 @@ struct simd_cast_fn<__m256i_t, __m128i_t> {
 template <>
 struct simd_cast_fn<int8_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(int8_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint32_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<int8_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<uint8_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(uint8_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint32_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint8_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m256i_t, int8_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int8_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint32_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, int8_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m256i_t, uint8_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint8_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint32_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, uint8_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<int16_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(int16_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint32_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<int16_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<uint16_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(uint16_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint32_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint16_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m256i_t, int16_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int16_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint32_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, int16_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<__m256i_t, uint16_t> {
     WJR_CONST WJR_INTRINSIC_INLINE uint16_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint32_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, uint16_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
 
 template <>
 struct simd_cast_fn<int32_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(int32_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint32_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<int32_t, __m128i_t>(v));
     }
 };
 
@@ -334,7 +334,7 @@ struct simd_cast_fn<uint32_t, __m256i_t> {
 template <>
 struct simd_cast_fn<__m256i_t, int32_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int32_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint32_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, int32_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
 
@@ -348,7 +348,7 @@ struct simd_cast_fn<__m256i_t, uint32_t> {
 template <>
 struct simd_cast_fn<int64_t, __m256i_t> {
     WJR_CONST WJR_INTRINSIC_INLINE __m256i operator()(int64_t v) const {
-        return simd_cast<__m128i_t, __m256i_t>(simd_cast<uint64_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, __m256i_t>(simd_cast<int64_t, __m128i_t>(v));
     }
 };
 
@@ -362,7 +362,7 @@ struct simd_cast_fn<uint64_t, __m256i_t> {
 template <>
 struct simd_cast_fn<__m256i_t, int64_t> {
     WJR_CONST WJR_INTRINSIC_INLINE int64_t operator()(__m256i v) const {
-        return simd_cast<__m128i_t, uint64_t>(simd_cast<__m256i_t, __m128i_t>(v));
+        return simd_cast<__m128i_t, int64_t>(simd_cast<__m256i_t, __m128i_t>(v));
     }
 };
